requirements.txt
django
django-storages
django-redis
django-filter
docker
djangorestframework
djangorestframework_simplejwt
djangorestframework-datatables
drf-extensions
mysqlclient
#psycopg2
boto3
django-prometheus
django-csp
django-cors-headers

#celery
#celery[redis]==5.2.2
django-celery-beat
#django-celery-results==2.2.0
gevent

celery[redis]
django-celery-results

sentry-sdk

pandas
xlrd
openpyxl
xlsxwriter
#fontawesome-free

rarfile
py7zr
crccheck

imap_tools

humanize
django-ckeditor
django-widget-tweaks

oauthlib>=3.0.0
django-lti-toolbox
PyLTI1p3

pillow

channels_redis
daphne

# HTML to text tool
beautifulsoup4

# Multi select with search
django-select2

# Canvas API
canvasapi
